requirements.txt
flask==3.0.0
python-dotenv==1.0.0
crewai==0.11.0
PyPDF2==3.0.1
pytesseract==0.3.10
pdf2image==1.16.3
python-multipart==0.0.6